Add auth controller, organize usings
