+++ conf/routes
@@ -16,9 +16,9 @@ GET     /bower_components/*file     controllers.Assets.at(path="/public/bower_co
 GET     /assets/*file               controllers.Assets.at(path="/public/dev/assets", file)
 
 # Users endpoints
-GET     /api/user/:id                   @controllers.UsersController.getUserById(id: String)
-PUT     /api/user/:id                   @controllers.UsersController.updateUserById(id: String)
+GET     /api/user/:id               @controllers.UsersController.getUserById(id: String)
+PUT     /api/user/:id               @controllers.UsersController.updateUserById(id: String)
 
 # Auth endpoints
-POST    /api/auth/create                @controllers.UsersController.createUser
-POST    /api/auth/login                 @controllers.UsersController.getByEmailAndPassword
+POST    /api/auth/create            @controllers.AuthController.createUser
+POST    /api/auth/login             @controllers.AuthController.login
